Refactor: Display에 Admin 출력 관련 메소드 AdminDisplay에 이동 / 메소드 이름 더욱 의미 있는 걸로 수정 / 선택 띄어쓰기 두번 -> 탭으로 변경 / Admin Menu 순서에 맞게 메소드 위치 수정
