+++ KioskProject/src/Display.java
@@ -9,13 +9,14 @@ import java.util.stream.Collectors;
 
 public class Display {
     // 필드
-    int password = 1004;
     static List<Order> orderList = new ArrayList<>();
     static List<Menu> menuList = new ArrayList<>();
     static Map<String, List<Product>> products = new HashMap<>();
 
     static Scanner scanner = new Scanner(System.in);
     static Order order = new Order();
+
+    static AdminDisplay adminDisplay = new AdminDisplay();
 
     static {
         // 데이터 초기화
@@ -60,14 +61,16 @@ public class Display {
 
         int number = Integer.parseInt(scanner.nextLine());
         if (number == 0) {
-            printLoginAdmin();
-        } else if (number >= 1 && number <= menuList.size()) {
+            adminDisplay.printAdminLogin();
+        }
+        else if (number >= 1 && number <= menuList.size()) {
             printProductList(number);
         } else if (number == menuList.size() + 1) {
             printBasket();
         } else if (number == menuList.size() + 2) {
             printCancelBasket();
-        } else {
+        }
+        else {
             System.out.println("잘못된 값을 입력했습니다. 메인으로 돌아갑니다.");
         }
     }
@@ -123,487 +126,4 @@ public class Display {
         }
     }
 
-    private void printProductList(int number) throws InterruptedException {
-
-        Menu menu = menuList.get(number - 1);
-        String menuName = menu.getName();
-
-        System.out.println("\n버커킹 에 오신걸 환영합니다.");
-        System.out.println("아래 상품메뉴판을 보시고 상품을 골라 입력해주세요\n");
-
-        System.out.println("[ " + menuName + " MENU ]");
-
-        List<Product> productList = products.get(menuName);
-        for (int i = 0; i < productList.size(); i++) {
-            System.out.println(
-                    i + 1 + ". " + productList.get(i).getName() + "\t | " + productList.get(i).getPrice()
-                            + "\t | " + productList.get(i).getExplain());
-        }
-
-        System.out.println("\n[ ORDER MENU ]");
-        System.out.println(productList.size() + 1 + ". 장바구니 확인");
-        System.out.println(productList.size() + 2 + ". 주문 취소\n");
-
-        int setOrder = Integer.parseInt(scanner.nextLine());
-        if (setOrder >= 1 && setOrder <= productList.size()) {
-            printAddBasket(productList.get(setOrder - 1));
-        } else if (setOrder == productList.size() + 1) {
-            printBasket();
-        } else if (setOrder == productList.size() + 2) {
-            printCancelBasket();
-        } else {
-            System.out.println("잘못된 값을 입력했습니다. 메인으로 돌아갑니다.");
-        }
-    }
-
-    private void printAddBasket(Product orderProduct) {
-
-        System.out.println("\n" +
-                orderProduct.getName() + " | " + orderProduct.getPrice() + " | " + orderProduct.getExplain());
-        System.out.println("위 메뉴를 장바구니에 추가하시겠습니까?");
-        System.out.println("1. 확인  2. 취소\n");
-
-        int orderSelect = Integer.parseInt(scanner.nextLine());
-
-        switch (orderSelect) {
-            case 1 -> {
-                order.addProduct(orderProduct);
-                System.out.println("\n" + orderProduct.getName() + "가 장바구니에 추가되었습니다.");
-            }
-            case 2 -> System.out.println("\n메인화면으로 돌아갑니다.");
-            default -> System.out.println("잘못된 값을 입력했습니다.");
-        }
-    }
-
-    private void printBasket() throws InterruptedException {
-
-        // 주문 내역 확인 및 토탈 가격 확인
-        System.out.println("\n[ Orders ]");
-        if (order.getBasket().isEmpty()) {
-            {
-                System.out.println("장바구니가 비어있습니다.");
-            }
-        } else {
-            List<Product> basket = order.getBasket();
-            for (Product product : basket) {
-                double price = product.getPrice();
-                String name = product.getName();
-                String explain = product.getExplain();
-                System.out.println(name + " | " + price + " | " + explain);
-            }
-        }
-        System.out.println("\n[ Total ]");
-        System.out.println(order.getBasketTotalPrice());
-        System.out.println("\n고객님의 주문 요청사항: " + order.getMessage() + "\n");
-        System.out.println("1. 주문하기  2.주문 요청사항 입력하기  3. 메뉴판\n");
-        int orderSelect = Integer.parseInt(scanner.nextLine());
-        if (orderSelect == 1) {
-            List<Product> basket = order.getBasket();
-            if (basket.isEmpty()) {
-                System.out.println("\n장바구니가 비어있어서 주문을 할 수 없습니다.");
-                printBasket();
-            } else {
-                printCompletedOrder();
-            }
-        }
-        if (orderSelect == 2) {
-            printOrderMessage();
-        }
-        if (orderSelect == 3) {
-            System.out.println("\n메인화면으로 돌아갑니다.");
-        }
-    }
-
-    private void printOrderMessage() throws InterruptedException {
-        System.out.println("\n요구사항을 입력하세요\n");
-        String message = scanner.nextLine();
-        if (!isLengthTwenty(message)) {
-            System.out.println("요구사항은 20자 이내 작성해주세요.");
-            printOrderMessage();
-        }
-        order.saveOrderMessage(message);
-        printBasket();
-    }
-
-    private boolean isLengthTwenty(String message) {
-        return message.length() <= 20;
-    }
-
-    private void printCancelBasket() throws InterruptedException {
-
-        System.out.println("\n진행하던 주문을 취소하시겠습니까?");
-        System.out.print("1. 확인  2. 취소\n");
-
-        int orderSelect = Integer.parseInt(scanner.nextLine());
-        switch (orderSelect) {
-            case 1 -> {
-                order = new Order();
-                System.out.println("\n진행하던 주문이 취소되었습니다.");
-            }
-            case 2 -> {
-                System.out.println("\n메인화면으로 돌아갑니다.");
-                printMain();
-            }
-        }
-    }
-
-    private void printCompletedOrder() throws InterruptedException {
-
-        System.out.println("\n주문이 완료되었습니다!\n");
-        order.setNumber();
-        order.setOrderTime();
-        orderList.add(order);
-        System.out.println("대기번호는 [ " + Order.orderNumber + " ] 번 입니다.");
-        order = new Order();
-        System.out.println("(3초후 메뉴판으로 돌아갑니다.)");
-        Thread.sleep(3000);
-    }
-
-    private void printLoginAdmin() throws InterruptedException {
-
-        System.out.println("\n[ ONLY ADMIN ]\n");
-        System.out.println("1. 비밀번호 입력  2.돌아가기\n");
-
-        int mainSelect = Integer.parseInt(scanner.nextLine());
-        switch (mainSelect) {
-            case 1 -> {
-                System.out.println("\n[ 비밀번호를 입력하세요 ]\n");
-                int passwordInput = Integer.parseInt(scanner.nextLine());
-                if (passwordInput == password) {
-                    printAdminMenu();
-                } else {
-                    System.out.println("\n비밀번호를 틀렸습니다.");
-                    printLoginAdmin();
-                }
-            }
-            case 2 -> System.out.println("\n메인화면으로 돌아갑니다.");
-            default -> {
-                System.out.println("잘못된 값을 입력했습니다.");
-                printLoginAdmin();
-            }
-        }
-    }
-
-    private void printAdminMenu() throws InterruptedException {
-
-        System.out.println("\n[ ADMIN MENU ]\n");
-        System.out.println("1. 대기 주문 목록 조회  2. 완료 주문 목록 조회 3. 메뉴 및 상품 추가  4. 메뉴 및 상품 삭제");
-        System.out.println("5. 총 판매 금액 현황 6. 총 판매 상품 현황 7. 비밀번호 변경  8. 돌아가기");
-
-        int mainSelect = Integer.parseInt(scanner.nextLine());
-        switch (mainSelect) {
-            case 1 -> printWaitingOrder();
-            case 2 -> printCompletedOrderList();
-            case 3 -> {
-                addMenu();
-                printAdminMenu();
-            }
-            case 4 -> printMenuAndProduct();
-            case 5 -> printSalesTotalPrice();
-            case 6 -> printSalesTotalProduct();
-            case 7 -> printChangeAdminPassword();
-            case 8 -> printMain();
-            default -> {
-                System.out.println("\n잘못된 값을 입력했습니다.");
-                printAdminMenu();
-            }
-        }
-    }
-
-    private void printWaitingOrder() throws InterruptedException {
-
-        System.out.println("\n[ 대기 주문 목록 ]\n");
-
-        List<Order> waitingOrderList = getWaitingOrderList();
-        for (Order currentOrder : waitingOrderList) {
-            System.out.println(currentOrder.getNumber() + "번 대기 주문 | 요구 사항: " + currentOrder.getMessage() + "\t | "
-                    + currentOrder.getOrderTime());
-            for (Product product : currentOrder.getBasket()) {
-                System.out.println("- " + product.getName());
-            }
-            System.out.println("--------------------------------------------------");
-
-        }
-
-        System.out.println("총 주문 금액");
-        System.out.println(getSaleWaitingTotalPrice());
-        System.out.println("\n1. 대기 주문 완료 처리 하기  2. ADMIN MENU로 돌아가기\n");
-
-        int input = Integer.parseInt(scanner.nextLine());
-        switch (input) {
-            case 1 -> waitingOrderInput();
-            case 2 -> printAdminMenu();
-            default -> {
-                System.out.println("잘못된 값을 입력했습니다.");
-                printWaitingOrder();
-            }
-        }
-    }
-
-    private void waitingOrderInput() throws InterruptedException {
-
-        System.out.println("\n완료 처리 할 대기 주문 번호 입력\n");
-
-        int input = Integer.parseInt(scanner.nextLine());
-
-        List<Order> waitingOrderList = getWaitingOrderList();
-
-        boolean isRemoved = false;
-
-        for (Order currentOrder : waitingOrderList) {
-            if (input == currentOrder.getNumber()) {
-                currentOrder.changeWaitingToCompleted();
-                System.out.println("\n" + input + "번 주문 완료 처리 되었습니다.\n");
-                isRemoved = true;
-                break;
-            }
-        }
-        if (!isRemoved) {
-            System.out.println("잘못된 값을 입력했습니다.");
-        }
-        printWaitingOrder();
-    }
-
-    private List<Order> getWaitingOrderList() {
-        return orderList.stream().filter(o -> o.getOrderStatus() == OrderStatus.WAITING).collect(Collectors.toList());
-    }
-
-    private void printCompletedOrderList() {
-
-        List<Order> completedOrderList = getCompletedOrderList();
-
-        System.out.println("\n[ 완료 주문 목록 ]\n");
-
-        for (Order currentOrder : completedOrderList) {
-            System.out.println(
-                    currentOrder.getNumber() + "번 완료 주문 | 요구 사항: " + currentOrder.getMessage() + "\t | "
-                            + currentOrder.getOrderTime()
-                            + "\t | " + currentOrder.getOrderCompleteTime());
-            for (Product product : currentOrder.getBasket()) {
-                System.out.println("- " + product.getName());
-            }
-            System.out.println("--------------------------------------------------");
-
-            System.out.println("총 주문 금액");
-            System.out.println(getSaleTotalPrice());
-        }
-    }
-
-    private List<Order> getCompletedOrderList() {
-        return orderList.stream().filter(o -> o.getOrderStatus() == OrderStatus.COMPLETED).collect(Collectors.toList());
-    }
-
-    private void printAddMenu() {
-
-        System.out.println("\n메뉴 이름을 입력해 주세요.");
-        String name = scanner.nextLine();
-
-        System.out.println("\n메뉴 설명을 입력해주세요.");
-        String explain = scanner.nextLine();
-
-        menuList.add(new Menu(name, explain));
-        products.put(name, new ArrayList<>());
-
-        System.out.println("\n메뉴 생성 완료!");
-    }
-
-    private void printAddProduct() {
-
-        System.out.println("\n메뉴를 선택해 주세요.");
-
-        for (int i = 0; i < menuList.size(); i++) {
-            System.out.println(i + 1 + ". " + menuList.get(i).getName());
-        }
-        int menuNum = Integer.parseInt(scanner.nextLine());
-        System.out.println("\n" + menuList.get(menuNum - 1).getName() + " 이(가) 선택 되었습니다.");
-        System.out.println("\n상품 이름을 입력해 주세요.");
-        String name = scanner.nextLine();
-
-        System.out.println("\n상품 설명을 입력해주세요.");
-        String explain = scanner.nextLine();
-
-        System.out.println("\n가격을 입력해주세요.");
-        double price = Double.parseDouble(scanner.nextLine());
-
-        String menuName = menuList.get(menuNum - 1).getName();
-        List<Product> newList = products.get(menuName);
-        newList.add(new Product(name, explain, price));
-
-        System.out.println("\n상품 생성 완료!");
-    }
-
-    private void printMenuAndProduct() {
-
-        System.out.println("1. 메뉴삭제 2. 상품삭제");
-        int select = Integer.parseInt(scanner.nextLine());
-        switch (select) {
-            case 1 -> printMenu();
-            case 2 -> printProduct();
-            default -> System.out.println("잘못된 입력입니다.");
-        }
-    }
-
-    private void deleteProduct() {
-
-        int delId = Integer.parseInt(scanner.nextLine());
-        boolean isRemoved = false;
-        for (Menu menu : menuList) {
-            String menuName = menu.getName();
-            List<Product> productList = Display.products.get(menuName);
-            for (int i = 0; i < productList.size(); i++) {
-                Product product = productList.get(i);
-                if (product.getId() == delId) {
-                    System.out.println("ID : " + product.getId() + "\t | " + product.getName()
-                            + "\t | " + product.getPrice() + "\t | " + product.getExplain()
-                            + "상품이 삭제되었습니다.");
-                    productList.remove(i);
-                    isRemoved = true;
-                    break;
-                }
-            }
-        }
-        if (!isRemoved) {
-            System.out.println("잘못된 ID 입니다.");
-        }
-    }
-
-    private void deleteMenu() {
-
-        int delId = Integer.parseInt(scanner.nextLine());
-        boolean isRemoved = false;
-        for (int i = 0; i < menuList.size(); i++) {
-            Menu menu = menuList.get(i);
-            if (menu.getId() == delId) {
-                System.out.println(
-                        "ID : " + menu.getId() + "\t | " + menu.getName()
-                                + "\t | " + menu.getExplain() + "메뉴가 삭제되었습니다.");
-                menuList.remove(i);
-                products.remove(menu.getName());
-                isRemoved = true;
-                break;
-            }
-        }
-        if (!isRemoved) {
-            System.out.println("잘못된 ID 입니다.");
-        }
-    }
-
-    private void printSalesTotalPrice() throws InterruptedException {
-
-        System.out.println("\n[ 총 판매 금액 현황 ]\n");
-        System.out.println("현재까지 총 판매된 금액은 [ " + getSaleTotalPrice() + " ] 입니다.\n");
-        System.out.println("1. ADMIN MENU로 돌아가기  2. 메인으로 돌아가기\n");
-
-        int mainSelect = Integer.parseInt(scanner.nextLine());
-        switch (mainSelect) {
-            case 1 -> printAdminMenu();
-            case 2 -> printMain();
-            default -> {
-                System.out.println("잘못된 값을 입력했습니다.");
-                printSalesTotalPrice();
-            }
-        }
-    }
-
-    private void printSalesTotalProduct() throws InterruptedException {
-
-        System.out.println("\n[ 총 판매상품 목록 현황 ]\n");
-        System.out.println("현재까지 총 판매된 상품 목록은 아래와 같습니다.\n");
-
-        for (Order currentOrder : orderList) {
-            for (Product product : currentOrder.getBasket()) {
-                System.out.println(
-                        "- " + product.getName() + "\t | " + product.getPrice());
-            }
-        }
-
-        System.out.println("\n1. ADMIN MENU로 돌아가기  2. 메인으로 돌아가기\n");
-        int mainSelect = Integer.parseInt(scanner.nextLine());
-        switch (mainSelect) {
-            case 1 -> printAdminMenu();
-            case 2 -> printMain();
-            default -> {
-                System.out.println("잘못된 값을 입력했습니다.");
-                printSalesTotalPrice();
-            }
-        }
-    }
-
-    private void addMenu() throws InterruptedException {
-
-        System.out.println("\n어떤 항목을 추가하시겠습니까?");
-        System.out.println("1.메뉴        2.상품       3.돌아가기");
-
-        int num = Integer.parseInt(scanner.nextLine());
-        switch (num) {
-            case 1 -> printAddMenu();
-            case 2 -> printAddProduct();
-            case 3 -> printAdminMenu();
-            default -> System.out.println("\n잘못된 입력입니다.");
-        }
-    }
-
-    private void printChangeAdminPassword() throws InterruptedException {
-
-        System.out.println("\n[ ADMIN 비밀번호 변경 ]\n");
-        System.out.print("현재 비밀번호 입력: ");
-
-        int inputPassword = Integer.parseInt(scanner.nextLine());
-        if (inputPassword == password) {
-            System.out.println("\n[ ADMIN 비밀번호 변경 ]\n");
-            System.out.print("변경할 비밀번호 입력: ");
-            password = Integer.parseInt(scanner.nextLine());
-            printAdminMenu();
-        } else {
-            System.out.println();
-            System.out.println("비밀번호를 틀렸습니다.");
-            printChangeAdminPassword();
-        }
-    }
-
-    private void printMenu() {
-
-        for (Menu value : menuList) {
-            System.out.println(
-                    "ID : " + value.getId() + "\t | " + value.getName()
-                            + "\t | " + value.getExplain());
-        }
-        System.out.println("삭제할 메뉴 ID: ");
-        deleteMenu();
-    }
-
-    private void printProduct() {
-
-        for (Menu menu : menuList) {
-            String menuName = menu.getName();
-            List<Product> productList = Display.products.get(menuName);
-            for (Product product : productList) {
-                System.out.println(
-                        "ID : " + product.getId() + "\t | " + product.getName()
-                                + "\t | " + product.getPrice() + "\t | " + product.getExplain());
-            }
-        }
-        System.out.println("삭제할 상품 ID: ");
-        deleteProduct();
-    }
-
-    private static double getSaleWaitingTotalPrice() {
-
-        double saleTotalPrice = 0;
-        for (Order order : orderList) {
-            if (order.getOrderStatus() == OrderStatus.WAITING) {
-                saleTotalPrice += order.getBasketTotalPrice();
-            }
-        }
-        return saleTotalPrice;
-    }
-
-    private static double getSaleTotalPrice() {
-
-        double saleTotalPrice = 0;
-        for (Order currentOrder : orderList) {
-            saleTotalPrice += currentOrder.getBasketTotalPrice();
-        }
-        return saleTotalPrice;
-    }
 }
